// src/types.ts
export interface ExperienceItem {
  title: string;
  company: string;
  start_date: string;
  end_date: string | null;
  description: string;
  current: boolean;
}

export interface EducationItem {
  institution: string;
  degree: string;
  field: string;
  start_date: string;
  end_date: string | null;
  current: boolean;
}